pyproject.toml
[project]
name = "spatial-microckg-agent"
version = "0.1.0"
description = "Disease-Agnostic Spatial-MicroCKG Agent: spatial multi-omics pipeline with Stabl feature selection, BioCypher Knowledge Graph, and LLM evidence traceability."
requires-python = ">=3.11"
dependencies = [
    "requests",
    "scanpy>=1.10",
    "squidpy>=1.5",
    "anndata",
    "pandas",
    "numpy",
    "scipy",
    "scikit-learn>=1.3.2,<1.7",
    "matplotlib",
    "seaborn",
    "biocypher>=0.5",
    "networkx",
    "langchain>=0.2",
    "langchain-google-genai",
    "langchain-openai",
    "python-dotenv",
    "tqdm",
    "joblib",
    "stabl @ git+https://github.com/gregbellan/Stabl.git@v1.0.1-lw",
    "igraph>=1.0.0",
    "leidenalg>=0.11.0",
    "langchain-ollama",
    "mygene",
    "gseapy",
    "chembl-webresource-client",
    "pyvis>=0.3.2",
]

[dependency-groups]
dev = [
    "jupyterlab",
    "ipykernel",
]
